akasha_data/chronojanus/gold-trader/requirements.txt
kafka-python
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
xgboost
ta
requests
python-dateutil
nltk
transformers
tensorflow
hmmlearn
scipy
statsmodels
lightgbm
